refactor: components refactoring: add folder smart, common in components folder
