+++ src/components/common/footer/Footer.scss
@@ -1,0 +1,1 @@
+@import '../../../style/utils/variable';
