+++ src/components/common/header/Header.scss
@@ -1,5 +1,4 @@
-@import '../../style/utils/variable';
-
+@import '../../../style/utils/variable';
 
 .header {
     width: 100%;
