+++ src/components/smart/cart-summary/CartSummary.module.scss
@@ -1,4 +1,4 @@
-@import "../../style/utils/variable";
+@import "../../../style/utils/variable";
 
 .cartSummary {
   display: flex;
