+++ src/components/smart/personal-data-form/PersonalDataForm.module.scss
@@ -1,4 +1,4 @@
-@import "../../style/utils/variable";
+@import "../../../style/utils/variable";
 
 .form {
   display: flex;
