+++ src/components/smart/personal-data-form/card-details/CardDetails.module.scss
