+++ src/components/smart/personal-data-form/form-group/FormGroup.module.scss
@@ -1,4 +1,4 @@
-@import "../../../style/utils/variable";
+@import "../../../../style/utils/variable";
 
 
 .formGroup {
