+++ src/containers/item-buy-modal/ItemBuyModal.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import styles from './ItemBuyModal.module.scss'
 import ReactDOM from 'react-dom'
-import {PersonalDataForm} from '../../components/personal-data-form/PersonalDataForm'
+import {PersonalDataForm} from '../../components/smart/personal-data-form/PersonalDataForm'
 
 
 interface IItemBuyModal {
